fetch events from db and display markers on map with title and desc

+++ app/src/main/java/com/example/apiapp/data/objects/EventAddressInformation.kt
@@ -3,6 +3,6 @@ package com.example.apiapp.data.objects
 data class EventAddressInformation(
     val address: String,
     val city: String,
-    val lat: Int,
-    val lng: Int
+    val lat: Float,
+    val lng: Float
 )
